feat: implement MosaicDashboardChart and related components for enhanced chart rendering
