+++ packages/mosaic/src/chart/MosaicDashboardChartRenderer.tsx
@@ -1,8 +1,6 @@
 import {SpinnerPane} from '@sqlrooms/ui';
 import {BarChart3Icon} from 'lucide-react';
-import React, {type FC, useCallback} from 'react';
-import {ChartSettingsPanel} from './chart-settings';
-import {MosaicDashboardPanelLayout} from '../dashboard/MosaicDashboardPanelLayout';
+import {type FC} from 'react';
 import {MosaicDashboardChartHeaderActions} from './MosaicDashboardChartHeaderActions';
 import {
   type MosaicDashboardPanelRenderer,
@@ -11,19 +9,13 @@ import {
   useStoreWithMosaicDashboard,
 } from '../dashboard/MosaicDashboardSlice';
 import {useChartTypeDefinition} from '../chart-types/useChartTypeDefinition';
-import {useChartRetainer} from './useChartRetainer';
-import {useBrushSelectionParams} from './useBrushSelectionParams';
-import {
-  ChartTypeDefinition,
-  isComponentChartType,
-  isSpecChartType,
-} from '../chart-types/base-types';
-import {MosaicDashboardVgPlotChart} from './MosaicDashboardVgPlotChart';
-import {useGenerateSpec, UseGenerateSpecResult} from './useGenerateSpec';
-import {MosaicReadyConnection} from '../MosaicSlice';
+import {MosaicDashboardChart} from './MosaicDashboardChart';
 
-const MosaicDashboardChartRenderer: FC<ChartPanelRendererProps> = (props) => {
-  const {panel} = props;
+const MosaicDashboardChartRenderer: FC<ChartPanelRendererProps> = ({
+  panel,
+  dashboardId,
+  selectionName,
+}) => {
   const connection = useStoreWithMosaicDashboard(
     (state) => state.mosaic.connection,
   );
@@ -60,8 +52,10 @@ const MosaicDashboardChartRenderer: FC<ChartPanelRendererProps> = (props) => {
   }
 
   return (
-    <MosaicDashboardChartContent
-      {...props}
+    <MosaicDashboardChart
+      dashboardId={dashboardId}
+      selectionName={selectionName}
+      panel={panel}
       chartTypeDef={chartTypeDef}
       tableName={tableName}
       connection={connection}
@@ -75,106 +69,3 @@ export const mosaicDashboardChartRenderer: MosaicDashboardPanelRenderer<ChartPan
     headerActions: MosaicDashboardChartHeaderActions,
     icon: BarChart3Icon,
   };
-
-type MosaicDashboardChartContentProps = ChartPanelRendererProps & {
-  chartTypeDef: ChartTypeDefinition;
-  tableName: string;
-  connection: MosaicReadyConnection;
-};
-
-const MosaicDashboardChartContent: FC<MosaicDashboardChartContentProps> = ({
-  chartTypeDef,
-  tableName,
-  connection,
-  ...props
-}) => {
-  const {dashboardId, panel} = props;
-
-  const updatePanel = useStoreWithMosaicDashboard(
-    (state) => state.mosaicDashboard.updatePanel,
-  );
-
-  const isSettingsOpen = panel.config.settingsOpen;
-
-  const handleOpenChange = useCallback(
-    (isOpen: boolean) => {
-      updatePanel(dashboardId, panel.id, {
-        config: {...panel.config, settingsOpen: isOpen},
-      });
-    },
-    [dashboardId, panel.config, panel.id, updatePanel],
-  );
-
-  const spec = useGenerateSpec(tableName, panel.config.settings, chartTypeDef);
-
-  const settingsContent = (
-    <ChartSettingsPanel
-      dashboardId={dashboardId}
-      panel={panel}
-      spec={spec.spec}
-      tableName={tableName}
-      onClose={() => handleOpenChange(false)}
-    />
-  );
-
-  const chartContent = (
-    <div className="h-full overflow-auto p-2">
-      <ChartContentRenderer
-        {...props}
-        chartTypeDefinition={chartTypeDef}
-        tableName={tableName}
-        connection={connection}
-        spec={spec}
-      />
-    </div>
-  );
-
-  return (
-    <div className="h-full min-h-0">
-      <MosaicDashboardPanelLayout
-        isOpen={isSettingsOpen}
-        onIsOpenChange={handleOpenChange}
-        settings={settingsContent}
-        content={chartContent}
-      />
-    </div>
-  );
-};
-
-type ChartContentRendererProps = ChartPanelRendererProps & {
-  chartTypeDefinition: ChartTypeDefinition;
-  tableName: string;
-  connection: MosaicReadyConnection;
-  spec: UseGenerateSpecResult;
-};
-
-const ChartContentRenderer: FC<ChartContentRendererProps> = (props) => {
-  const {
-    selectionName,
-    panel,
-    dashboardId,
-    chartTypeDefinition,
-    tableName,
-    connection,
-    spec,
-  } = props;
-
-  const retention = useChartRetainer(dashboardId, panel.id);
-  const params = useBrushSelectionParams(selectionName);
-
-  if (isSpecChartType(chartTypeDefinition)) {
-    return <MosaicDashboardVgPlotChart {...props} spec={spec} />;
-  }
-
-  if (isComponentChartType(chartTypeDefinition)) {
-    return React.createElement(chartTypeDefinition.renderer, {
-      tableName,
-      config: panel.config,
-      coordinator: connection.coordinator,
-      params,
-      retention,
-    });
-  }
-
-  throw new Error('Unsupported chart type definition');
-};
